+++ TechnosoftProject/ViewModels/MenuDTO.cs
@@ -5,7 +5,7 @@ namespace Technosoft_Project.ViewModels
 {
     public class MenuDTO
     {
-        public int ID { get; set; }
+        public short ID { get; set; }
 
         [Required(ErrorMessage = "Menü ismi girmelisiniz.")]
         [Display(Name = "Menü Adı")]
